+++ src/finance.rs
@@ -1,0 +1,27 @@
+//! General functions related to finance.
+use crate::units::{Dimensionless, MoneyPerCapacity};
+
+/// Calculates the capital recovery factor (CRF) for a given lifetime and discount rate.
+///
+/// The CRF is used to annualize capital costs over the lifetime of an asset.
+pub fn capital_recovery_factor(lifetime: u32, discount_rate: Dimensionless) -> Dimensionless {
+    if lifetime == 0 {
+        return Dimensionless(0.0);
+    }
+    if discount_rate == Dimensionless(0.0) {
+        return Dimensionless(1.0) / Dimensionless(lifetime as f64);
+    }
+    let factor = (Dimensionless(1.0) + discount_rate).powi(lifetime as i32);
+    (discount_rate * factor) / (factor - Dimensionless(1.0))
+}
+
+/// Calculates the annual capital cost for a technology per unit of capacity
+pub fn annual_capital_cost(
+    capital_cost: MoneyPerCapacity,
+    lifetime: u32,
+    discount_rate: Dimensionless,
+) -> MoneyPerCapacity {
+    let crf = capital_recovery_factor(lifetime, discount_rate);
+    let total_capital_cost = capital_cost * crf;
+    total_capital_cost * crf
+}
