+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod agent;
 pub mod asset;
 pub mod commands;
 pub mod commodity;
+pub mod finance;
 pub mod id;
 pub mod input;
 pub mod log;
